Moved @tryghost/email-address to core folder
fix https://linear.app/ghost/issue/ENG-2402/email-addresses

- this commit moves the lib code and tests to the core folder so we can
  colocate code in one place rather than splitting it out across
  packages, which increases the cognitive load and overhead
