+++ ghost/core/core/server/services/email-address/EmailAddressParser.js.d.ts
@@ -1,0 +1,13 @@
+/* eslint-disable ghost/filenames/match-exported-class */
+
+export interface EmailAddress {
+    address: string;
+    name?: string;
+}
+
+declare class EmailAddressParser {
+    static parse(email: string): EmailAddress | null;
+    static stringify(email: EmailAddress): string;
+}
+
+export default EmailAddressParser;
