+++ ghost/core/core/server/services/email-service/EmailRenderer.js
@@ -8,7 +8,7 @@ const {isUnsplashImage} = require('@tryghost/kg-default-cards/lib/utils');
 const {textColorForBackgroundColor, darkenToContrastThreshold} = require('@tryghost/color-utils');
 const {DateTime} = require('luxon');
 const htmlToPlaintext = require('@tryghost/html-to-plaintext');
-const {EmailAddressParser} = require('@tryghost/email-addresses');
+const EmailAddressParser = require('../email-address/EmailAddressParser');
 const {registerHelpers} = require('./helpers/register-helpers');
 const crypto = require('crypto');
