+++ ghost/core/package.json
@@ -77,7 +77,6 @@
     "@tryghost/database-info": "0.3.29",
     "@tryghost/debug": "0.1.34",
     "@tryghost/domain-events": "1.0.0",
-    "@tryghost/email-addresses": "0.0.0",
     "@tryghost/email-mock-receiver": "0.3.10",
     "@tryghost/errors": "1.3.7",
     "@tryghost/helpers": "1.1.93",
@@ -193,6 +192,7 @@
     "nconf": "0.13.0",
     "node-fetch": "2.7.0",
     "node-jose": "2.2.0",
+    "nodemailer": "6.10.1",
     "otplib": "12.0.1",
     "papaparse": "5.3.2",
     "path-match": "1.2.4",
@@ -227,6 +227,7 @@
     "@types/jsonwebtoken": "9.0.9",
     "@types/node": "22.15.8",
     "@types/node-jose": "1.1.13",
+    "@types/nodemailer": "6.4.17",
     "@types/sinon": "17.0.4",
     "@types/supertest": "6.0.3",
     "c8": "8.0.1",
