+++ ghost/core/test/unit/server/services/newsletters/service.test.js
@@ -8,7 +8,7 @@ const mail = require('../../../../../core/server/services/mail');
 // Mocked utilities
 const urlUtils = require('../../../../utils/urlUtils');
 const {mockManager} = require('../../../../utils/e2e-framework');
-const {EmailAddressService} = require('@tryghost/email-addresses');
+const {EmailAddressService} = require('../../../../../core/server/services/email-address/EmailAddressService');
 const NewslettersService = require('../../../../../core/server/services/newsletters/NewslettersService');
 
 class TestTokenProvider {
