--- ghost/email-addresses/src/index.ts
@@ -1,2 +1,0 @@
-export * from './EmailAddressService';
-export * from './EmailAddressParser';
