--- ghost/email-addresses/src/libraries.d.ts
@@ -1,3 +1,0 @@
-declare module '@tryghost/errors';
-declare module '@tryghost/tpl';
-declare module '@tryghost/logging';
